REFACTOR: Replace variant with vector<uint8_t>

+++ src/image_factory.cpp
@@ -153,15 +153,15 @@ RawImage ImageFactory::create_image_rgb_interleaved_10bit() {
   RawImage image(m_width, m_height);
   const uint32_t band_count = 3; // RGB
   uint64_t pixel_count = m_width * m_height * band_count;
-  vector<uint16_t> pixels;
+  vector<uint8_t> pixels;
   pixels.reserve(pixel_count);
 
   // Fill Pixels
   for (uint32_t y = 0; y < m_height; y++) {
     for (uint32_t x = 0; x < m_width; x++) {
-      pixels.push_back(static_cast<uint16_t>(m_color_1)); // R or Y
-      pixels.push_back(static_cast<uint16_t>(m_color_2)); // G or U
-      pixels.push_back(static_cast<uint16_t>(m_color_3)); // B or V
+      pixels.push_back(static_cast<uint8_t>(m_color_1)); // R or Y
+      pixels.push_back(static_cast<uint8_t>(m_color_2)); // G or U
+      pixels.push_back(static_cast<uint8_t>(m_color_3)); // B or V
     }
   }
